Use the promethus common library
