+++ cmd/po-lint/main.go
@@ -24,14 +24,14 @@ import (
 	"github.com/ghodss/yaml"
 	v1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	v1alpha1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1alpha1"
-	"github.com/prometheus-operator/prometheus-operator/pkg/version"
+	"github.com/prometheus-operator/prometheus-operator/pkg/versionutil"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func main() {
-	version.RegisterParseFlags()
-	if version.ShouldPrintVersion() {
-		version.Print(os.Stdout, "po-lint")
+	versionutil.RegisterParseFlags()
+	if versionutil.ShouldPrintVersion() {
+		versionutil.Print(os.Stdout, "po-lint")
 		os.Exit(0)
 	}
